--remove interface --add all args constructor
